fix: Issues foud during testing
Issue 1 - Cities not identified correctly by using city_rejects table
Issue 2 - Giftbox and other page types incorrectly identified as cities
Issue 3 - Updated thrown error text

+++ app/SQLiteConnection.php
@@ -30,7 +30,9 @@ class SQLiteConnection {
      * Method uses the config file to find location of database and deletes it
      */
      public function deleteDatabase() {
-       unlink(Config::PATH_TO_SQLITE_FILE);
+       if (Config::PATH_TO_SQLITE_FILE) {
+         unlink(Config::PATH_TO_SQLITE_FILE);
+       }
      }
 
 }
